pyproject.toml
[project]
name = "mail-mcp-server"
version = "0.1.0"
description = "Email sending service using the FastMCP library."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=1.0.0",
    "python-dotenv>=0.19.0",
]

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

# 非包结构应用：仅同步依赖到 .venv，不把当前目录当作可安装包
[tool.uv]
package = false
